requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
pydantic
llama-cloud-services
llama-index-core
openai
numpy
pandas
chromadb